requirements.txt
Flask
Flask-Cors
Flask-WTF
WTForms
numpy
pandas
scikit-learn
xgboost
tensorflow
beautifulsoup4
requests
google-cloud-bigquery
pyarrow
seaborn
